requirements.txt
requests
python-dotenv
Flask
paddleocr
pillow